elastestbrowsers: CORRECTNESS REVIEW (read description)
This commit brings a full review over all Dockerfiles and scripts. It
makes source code safer (through extra checks for errors), easier to
read, and more idiomatic to the language they are written in.

However, NO FUNCTIONALITY CHANGES ARE INTRODUCED.

Changes in Dockerfiles:

* Add correct cleanup after all RUN commands:
- "apt-get install" should always be followed by
    "rm -rf /var/lib/apt/lists/*"
- "pip install" should always use "--no-cache-dir"

* Document why each package is installed. Otherwise, future devs won't
know if a package is still needed or not, and this ends up with the
usual: fat images that nobody knows if can be made slimmer.

Changes in scripts:

* Enforce "safe mode" everywhere (Bash options for strict error
checking)

This is done by enabling all checks that come disabled by default:
    set -o errexit -o errtrace -o pipefail -o nounset

Just testing locally, this has already found several errors that have
also been solved in this commit.

* Add comments and separate code in clearly visible sections.

* Simplify unnecessarily complex commands.

For example, in base/start-video-recording.sh:
Instead of:    ps ax | grep [f]fmpeg | awk '{ print $1'}
just do this!: pgrep ffmpeg