elastestbrowsers: disable PulseAudio autospawn
PulseAudio has an "autospawn" feature that will restart the daemon under
some circumstances, such as when the process crashes.

We want to have total control of when PulseAudio is started, and for how
long. If it crashes, we want to know and get an error in the logs.

This commit disables the PulseAudio autospawn feature, so it's
guaranteed that the daemon starts only through the entrypoint script.