elastestbrowsers: use a lock file to guard entrypoint execution
Creating a lock file allows the Java test application to check if the
entrypoint script has finished running.

Without this, we are in a difficult situation where tests might get run
by the Java application even before the required services are started.
This was especially obvious with the recording process, which sometimes
was being started before PulseAudio was running.

In general terms, Docker is missing an official mechanism that allows
waiting for the entrypoint script, so we have to use tricks like this
lock file as a "workaround".